# PrecastTracker.WebApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "PrecastTracker.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/ConcreteReport/csv": {
      "get": {
        "tags": [
          "ConcreteReport"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ConcreteReport": {
      "get": {
        "tags": [
          "ConcreteReport"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tester-report/tests-due-today": {
      "get": {
        "tags": [
          "TesterReport"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tester-report/tests-overdue": {
      "get": {
        "tags": [
          "TesterReport"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tester-report/tests-upcoming": {
      "get": {
        "tags": [
          "TesterReport"
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCylinderQueueResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tester-report/untested-placements": {
      "get": {
        "tags": [
          "TesterReport"
        ],
        "parameters": [
          {
            "name": "daysBack",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UntestedPlacementResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UntestedPlacementResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UntestedPlacementResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tester-report/test-set-day/{testSetDayId}": {
      "get": {
        "tags": [
          "TesterReport"
        ],
        "parameters": [
          {
            "name": "testSetDayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTestSetDayDetailsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTestSetDayDetailsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTestSetDayDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tester-report/test-set-day": {
      "post": {
        "tags": [
          "TesterReport"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestSetDayDataRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestSetDayDataRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestSetDayDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GetTestSetDayDetailsResponse": {
        "type": "object",
        "properties": {
          "testSetDayId": {
            "type": "integer",
            "format": "int32"
          },
          "dayNum": {
            "type": "integer",
            "format": "int32"
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "dateDue": {
            "type": "string",
            "format": "date-time"
          },
          "dateTested": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "jobCode": {
            "type": "string",
            "nullable": true
          },
          "jobName": {
            "type": "string",
            "nullable": true
          },
          "mixDesignCode": {
            "type": "string",
            "nullable": true
          },
          "requiredPsi": {
            "type": "integer",
            "format": "int32"
          },
          "pieceType": {
            "type": "string",
            "nullable": true
          },
          "castDate": {
            "type": "string",
            "format": "date-time"
          },
          "castTime": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "testCylinders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCylinderBreakDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "SaveTestSetDayDataRequest": {
        "type": "object",
        "properties": {
          "testSetDayId": {
            "type": "integer",
            "format": "int32"
          },
          "dateTested": {
            "type": "string",
            "format": "date-time"
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "cylinderBreaks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCylinderBreakInput"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TestCylinderBreakDto": {
        "type": "object",
        "properties": {
          "testCylinderId": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "breakPsi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TestCylinderBreakInput": {
        "type": "object",
        "properties": {
          "testCylinderId": {
            "type": "integer",
            "format": "int32"
          },
          "breakPsi": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TestCylinderQueueResponse": {
        "type": "object",
        "properties": {
          "testCylinderCode": {
            "type": "string",
            "nullable": true
          },
          "ovenId": {
            "type": "string",
            "nullable": true
          },
          "dayNum": {
            "type": "integer",
            "format": "int32"
          },
          "castDate": {
            "type": "string",
            "format": "date-time"
          },
          "castTime": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "jobCode": {
            "type": "string",
            "nullable": true
          },
          "jobName": {
            "type": "string",
            "nullable": true
          },
          "mixDesignCode": {
            "type": "string",
            "nullable": true
          },
          "requiredPsi": {
            "type": "integer",
            "format": "int32"
          },
          "pieceType": {
            "type": "string",
            "nullable": true
          },
          "testSetId": {
            "type": "integer",
            "format": "int32"
          },
          "testSetDayId": {
            "type": "integer",
            "format": "int32"
          },
          "dateDue": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UntestedPlacementResponse": {
        "type": "object",
        "properties": {
          "pourId": {
            "type": "integer",
            "format": "int32"
          },
          "placementId": {
            "type": "integer",
            "format": "int32"
          },
          "castDate": {
            "type": "string",
            "format": "date-time"
          },
          "castTime": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "jobCode": {
            "type": "string",
            "nullable": true
          },
          "jobName": {
            "type": "string",
            "nullable": true
          },
          "mixDesignCode": {
            "type": "string",
            "nullable": true
          },
          "pieceType": {
            "type": "string",
            "nullable": true
          },
          "volume": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}